SDSTOR-6954 : TestLogStore-Spdk crashes during shutdown (#236)
link: https://jirap.corp.ebay.com/browse/SDSTOR-6954

Address sanitizer error in the test. The reason is that during shutdown, stop_spdk() is called to free mempool in m_iomgr_internal_pools but the array elements is not reset as nullptr. For the second calls of  create_mempool for the fake restart, the pointer of m_iomgr_internal_pools elements are not nullptr but their contents are already freed.

* change :
 call m_iomgr_internal_pools.fill(nullptr) to set all elements to be nullptr